+++ src/main/java/babysnoozer/events/SetSnoozingEndPosEvent.java
@@ -4,13 +4,13 @@ package babysnoozer.events;
  * Created by Alexander Bischof on 12.01.15.
  */
 public class SetSnoozingEndPosEvent {
-  private final long endPos;
+  private final int endPos;
 
-  public SetSnoozingEndPosEvent(long endPos) {
+  public SetSnoozingEndPosEvent(int endPos) {
 	this.endPos = endPos;
   }
 
-  public long getEndPos() {
+  public int getEndPos() {
 	return endPos;
   }
 }
