+++ src/main/java/babysnoozer/handlers/RopeDistance.java
@@ -4,26 +4,26 @@ package babysnoozer.handlers;
  * Created by Alexander Bischof on 30.01.15.
  */
 public class RopeDistance {
-    private long startPos;
-    private long endPos;
+    private int startPos;
+    private int endPos;
 
-    public long getStartPos() {
+    public int getStartPos() {
         return startPos;
     }
 
-    public void setStartPos(long startPos) {
+    public void setStartPos(int startPos) {
         this.startPos = startPos;
     }
 
-    public long getEndPos() {
+    public int getEndPos() {
         return endPos;
     }
 
-    public void setEndPos(long endPos) {
+    public void setEndPos(int endPos) {
         this.endPos = endPos;
     }
 
-    public long getPosBy(int percentage){
+    public int getPosBy(int percentage){
         if (percentage < 0 || percentage > 100){
             throw new IllegalArgumentException("percentage between 0 and 100 expected");
         }
@@ -31,6 +31,6 @@ public class RopeDistance {
         long distance = Math.abs(endPos-startPos);
         distance = percentage * distance /100;
 
-        return (Math.min(endPos, startPos) + distance);
+        return (int)(Math.min(endPos, startPos) + distance);
     }
 }
