+++ src/main/java/babysnoozer/handlers/commands/PositionCommand.java
@@ -8,12 +8,12 @@ import static babysnoozer.tinkerforge.BrickStepperWrapper.*;
  * Created by Alexander Bischof on 10.03.15.
  */
 public class PositionCommand implements Command {
-  private final long position;
+  private final int position;
   private final Velocity velocity;
   private final Acceleration acceleration;
   private final Acceleration deacceleration;
 
-  public PositionCommand(long position, Velocity velocity,
+  public PositionCommand(int position, Velocity velocity,
                          Acceleration acceleration,
                          Acceleration deacceleration) {
 	this.position = position;
@@ -22,7 +22,7 @@ public class PositionCommand implements Command {
     this.deacceleration = deacceleration;
   }
 
-  public long getPosition() {
+  public int getPosition() {
 	return position;
   }
